Made a class for embedding
